+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-## TBD
+## 1.9.0 (2021-02-10)
 
 ### Enhancements